pyproject.toml
[tool.poetry]
name = "rnaseq_viz"
version = "0.0.1"
authors = ["Your email"]
readme = "README.md"
description = "See README.md"
license = "GNU General Public License v3.0"
packages = [
    { include = "rnaseq_viz" },
]

[tool.poetry.dependencies]
python = "3.10.11"
streamlit = "^1.37.1"
fastapi = {extras = ["all"], version = "^0.112.1"}
uvicorn = "^0.30.6"
boto3 = "^1.34.162"
pandas = "^2.2.2"
pyarrow = "^17.0.0"
pydantic = "^2.8.2"
pydantic-core = ">=2.0.0"
annotated-types = "^0.7.0"
h11 = "^0.14.0"
python-dotenv = "^1.0.1"
matplotlib = "^3.9.2"
seaborn = "^0.13.2"

[tool.poetry.dev-dependencies]
coverage = "~5.4"
flake8 = "3.8.4"
mkdocs = "1.3.0"
safety = "1.10.3"
twine = "3.3.0"
mkdocstrings = {extras = ["python"], version = "0.19.0"}
pytkdocs = {version = "0.16.1"}
pytest = "~7.2"

[[tool.poetry.source]]
name = "PyPI"
priority = "supplemental"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
